Extract Prompt Studio wheel helpers
